FIX: Various improvements to Object3D (#5668)
