+++ src/ansys/aedt/core/modeler/cad/object_3d.py
@@ -34,12 +34,17 @@ objects (points, lines, sheets, and solids) within the AEDT 3D Modeler.
 """
 from __future__ import absolute_import  # noreorder
 
+import math
 import os
 import re
 
+from ansys.aedt.core.application.variables import decompose_variable_value
 from ansys.aedt.core.generic.checks import min_aedt_version
 from ansys.aedt.core.generic.constants import AEDT_UNITS
+from ansys.aedt.core.generic.constants import PLANE
+from ansys.aedt.core.generic.constants import unit_converter
 from ansys.aedt.core.generic.errors import AEDTRuntimeError
+from ansys.aedt.core.generic.general_methods import _dim_arg
 from ansys.aedt.core.generic.general_methods import _to_boolean
 from ansys.aedt.core.generic.general_methods import _uname
 from ansys.aedt.core.generic.general_methods import clamp
@@ -51,6 +56,7 @@ from ansys.aedt.core.modeler.cad.elements_3d import BinaryTreeNode
 from ansys.aedt.core.modeler.cad.elements_3d import EdgePrimitive
 from ansys.aedt.core.modeler.cad.elements_3d import FacePrimitive
 from ansys.aedt.core.modeler.cad.elements_3d import VertexPrimitive
+from ansys.aedt.core.modeler.geometry_operators import GeometryOperators
 
 
 class Object3d(object):
@@ -78,6 +84,7 @@ class Object3d(object):
 
     def __init__(self, primitives, name=None):
         self._id = None
+        self._positions = None
         if name:
             self._m_name = name
         else:
@@ -102,6 +109,27 @@ class Object3d(object):
         self._volume = 0.0
         self._faces = []
         self._face_ids = []
+        self._is_polyline = None
+
+    @property
+    def is_polyline(self):
+        """Get or set if the body is originated by a polyline.
+
+        Returns
+        -------
+        bool
+        """
+        if self._is_polyline is None:
+            hist = self.history()
+            if hist.children and (hist.command == "CreatePolyline" or hist.properties.get("Segment Type", None)):
+                self._is_polyline = True
+            else:
+                self._is_polyline = False
+        return self._is_polyline
+
+    @is_polyline.setter
+    def is_polyline(self, value):
+        self._is_polyline = value
 
     @pyaedt_function_handler()
     def _bounding_box_unmodel(self):
@@ -1762,6 +1790,7 @@ class Object3d(object):
         if not success:
             raise AEDTRuntimeError(f"Could not clone the object {self.name}")
         new_name = new_obj_tuple[1][0]
+        self._primitives[new_name].is_polyline = self.is_polyline
         return self._primitives[new_name]
 
     @pyaedt_function_handler()
@@ -2051,3 +2080,990 @@ class Object3d(object):
             self.logger.error("Operation Failed generating Unclassified object. Check and retry")
             return False
         return True
+
+    @property
+    def start_point(self):
+        """Get the starting point in the polyline object.
+
+        This is a list of the ``[x, y, z]`` coordinates for the starting point in the polyline
+        object in the object's coordinate system
+
+        Returns
+        -------
+        list
+            List of the ``[x, y, z]`` coordinates for the starting point in the polyline
+            object.
+        """
+        try:
+            return self.points[0]
+        except (TypeError, IndexError):
+            return
+
+    @property
+    def end_point(self):
+        """List of the ``[x, y, z]`` coordinates for the ending point in the polyline
+        object in the object's coordinate system.
+
+        Returns
+        -------
+        list
+            List of the ``[x, y, z]`` coordinates for the ending point in the polyline
+            object.
+
+        References
+        ----------
+        >>> oEditor.GetVertexIDsFromObject
+        >>> oEditor.GetVertexPosition
+
+        """
+        try:
+            return self.points[-1]
+        except (TypeError, IndexError):
+            return
+
+    def _update_segments_and_points(self):
+        """Update the self._segment_types and the self._positions from the history.
+
+        This internal method is called by properties ``points`` and ``segment_types``.
+        It will be called only once after opening a new project, then the internal
+        variables are maintained updated.
+        It is a single update call for both properties because they are very similar,
+        and we can access to the history only once.
+        """
+
+        def _convert_points(p_in, dest_unit):
+            p_out = []
+            for ip in p_in:
+                v, u = decompose_variable_value(ip)
+                if u == "":
+                    p_out.append(v)
+                else:
+                    p_out.append(unit_converter(v, unit_system="Length", input_units=u, output_units=dest_unit))
+            return p_out
+
+        segments = []
+        points = []
+        try:
+            history = self.history()
+            h_segments = history.segments
+        except Exception:  # pragma: no cover
+            history = None
+            h_segments = None
+        if h_segments:
+            for i, c in enumerate(h_segments.values()):
+                # evaluate the number of points in the segment
+                attrb = list(c.properties.keys())
+                n_points = 0
+                for j in range(1, len(attrb) + 1):
+                    if "Point" + str(j) in attrb:
+                        n_points += 1
+                # get the segment type
+                s_type = c.properties["Segment Type"]
+                if i == 0:  # append the first point only for the first segment
+                    if s_type != "Center Point Arc":
+                        p = [
+                            c.properties["Point1/X"],
+                            c.properties["Point1/Y"],
+                            c.properties["Point1/Z"],
+                        ]
+                        points.append(_convert_points(p, self._primitives.model_units))
+                    else:
+                        p = [
+                            c.properties["Start Point/X"],
+                            c.properties["Start Point/Y"],
+                            c.properties["Start Point/Z"],
+                        ]
+                        points.append(_convert_points(p, self._primitives.model_units))
+                if s_type == "Line":
+                    segments.append(PolylineSegment("Line"))
+                    p = [
+                        c.properties["Point2/X"],
+                        c.properties["Point2/Y"],
+                        c.properties["Point2/Z"],
+                    ]
+                    points.append(_convert_points(p, self._primitives.model_units))
+                elif s_type == "3 Point Arc":
+                    segments.append(PolylineSegment("Arc"))
+                    p2 = [
+                        c.properties["Point2/X"],
+                        c.properties["Point2/Y"],
+                        c.properties["Point2/Z"],
+                    ]
+                    p3 = [
+                        c.properties["Point3/X"],
+                        c.properties["Point3/Y"],
+                        c.properties["Point3/Z"],
+                    ]
+
+                    points.append(_convert_points(p2, self._primitives.model_units))
+                    points.append(_convert_points(p3, self._primitives.model_units))
+                elif s_type == "Spline":
+                    segments.append(PolylineSegment("Spline", num_points=n_points))
+                    for p in range(2, n_points + 1):
+                        point_attr = "Point" + str(p)
+                        p2 = [
+                            c.properties[f"{point_attr}/X"],
+                            c.properties[f"{point_attr}/Y"],
+                            c.properties[f"{point_attr}/Z"],
+                        ]
+
+                        points.append(_convert_points(p2, self._primitives.model_units))
+                elif s_type == "Center Point Arc":
+                    p2 = [
+                        c.properties["Start Point/X"],
+                        c.properties["Start Point/Y"],
+                        c.properties["Start Point/Z"],
+                    ]
+                    p3 = [
+                        c.properties["Center Point/X"],
+                        c.properties["Center Point/Y"],
+                        c.properties["Center Point/Z"],
+                    ]
+                    start = _convert_points(p2, self._primitives.model_units)
+                    center = _convert_points(p3, self._primitives.model_units)
+                    plane = c.properties["Plane"]
+                    angle = c.properties["Angle"]
+                    arc_seg = PolylineSegment("AngularArc", arc_angle=angle, arc_center=center, arc_plane=plane)
+                    segments.append(arc_seg)
+                    self._evaluate_arc_angle_extra_points(arc_seg, start)
+                    points.extend(arc_seg.extra_points[:])
+
+        # perform validation
+        if history:
+            nn_segments = int(history.properties["Number of curves"])
+            nn_points = int(history.properties["Number of points"])
+        else:
+            nn_segments = None
+            nn_points = None
+        if not len(segments) == nn_segments:
+            raise AEDTRuntimeError("Failed to get the polyline segments from AEDT.")
+        if not len(points) == nn_points:
+            raise AEDTRuntimeError("Failed to get the polyline points from AEDT.")
+        # if succeeded save the result
+        self._segment_types = segments
+        self._positions = points
+
+    @property
+    def points(self):
+        """Polyline Points."""
+        if not self.is_polyline:
+            return
+        if self._positions:
+            return self._positions
+        else:
+            self._update_segments_and_points()
+            return self._positions
+
+    @property
+    def segment_types(self):
+        """List of the segment types of the polyline."""
+        if not self.is_polyline:
+            return
+        if self._segment_types:
+            return self._segment_types
+        else:
+            self._update_segments_and_points()
+            return self._segment_types
+
+    @property
+    def vertex_positions(self):
+        """List of the ``[x, y, z]`` coordinates for all vertex positions in the
+        polyline object in the object's coordinate system.
+
+        Returns
+        -------
+        list
+            List of the ``[x, y, z]`` coordinates for all vertex positions in the
+            polyline object.
+
+        References
+        ----------
+        >>> oEditor.GetVertexIDsFromObject
+        >>> oEditor.GetVertexPosition
+
+        """
+        if not self.is_polyline:
+            return
+        id_list = self._primitives.get_object_vertices(assignment=self.id)
+        position_list = [self._primitives.get_vertex_position(id) for id in id_list]
+        return position_list
+
+    @pyaedt_function_handler()
+    def _pl_point(self, pt):
+        pt_data = ["NAME:PLPoint"]
+        pt_data.append("X:=")
+        pt_data.append(_dim_arg(pt[0], self._primitives.model_units))
+        pt_data.append("Y:=")
+        pt_data.append(_dim_arg(pt[1], self._primitives.model_units))
+        pt_data.append("Z:=")
+        pt_data.append(_dim_arg(pt[2], self._primitives.model_units))
+        return pt_data
+
+    @pyaedt_function_handler()
+    def _point_segment_string_array(self):
+        """Retrieve the parameter arrays for specifying the points and segments of a polyline
+        used in the :class:`ansys.aedt.core.modeler.cad.primitives.Polyline` constructor.
+
+        Returns
+        -------
+        list
+        """
+        position_list = self.points
+        segment_types = self.segment_types
+
+        # Add a closing point if needed
+        arg_1 = [
+            "NAME:PolylineParameters",
+            "IsPolylineCovered:=",
+            self._is_covered,
+            "IsPolylineClosed:=",
+            self._is_closed,
+        ]
+
+        # PointsArray
+        points_str = ["NAME:PolylinePoints"]
+        points_str.append(self._pl_point(position_list[0]))
+
+        # Segments Array
+        segment_str = ["NAME:PolylineSegments"]
+
+        pos_count = 0
+        vertex_count = 0
+        index_count = 0
+
+        while vertex_count <= len(segment_types):
+            try:
+                current_segment = None
+                if vertex_count == len(segment_types):
+                    if self._is_closed:
+                        # Check the special case of a closed polyline needing an additional Line segment
+                        if position_list[0] != position_list[-1]:
+                            position_list.append(position_list[0])
+                            current_segment = PolylineSegment("Line")
+                    else:
+                        break
+                else:
+                    current_segment = segment_types[vertex_count]
+            except Exception:
+                raise IndexError("Number of segments inconsistent with the number of points!")
+
+            if current_segment:
+                seg_str = self._segment_array(
+                    current_segment, start_index=index_count, start_point=position_list[pos_count]
+                )
+                segment_str.append(seg_str)
+
+                pos_count_incr = 0
+                for i in range(1, current_segment.num_points):
+                    if current_segment.type == "AngularArc":
+                        points_str.append(self._pl_point(current_segment.extra_points[i - 1]))
+                        index_count += 1
+                        pos_count_incr += 1
+                    else:
+                        if (pos_count + i) == len(position_list):
+                            if current_segment.type == "Arc" and self._is_closed:
+                                position_list.append(position_list[0])
+                            else:
+                                err_str = "Insufficient points in position_list to complete the specified segment list"
+                                raise ValueError(err_str)
+                        points_str.append(self._pl_point(position_list[pos_count + i]))
+                        pos_count_incr += 1
+                        index_count += 1
+                pos_count += pos_count_incr
+                vertex_count += 1
+            else:
+                break
+
+        arg_1.append(points_str)
+        arg_1.append(segment_str)
+
+        # Poly Line Cross Section
+        arg_1.append(self._xsection)
+
+        return arg_1
+
+    @pyaedt_function_handler()
+    def _evaluate_arc_angle_extra_points(self, segment, start_point):
+        """Evaluate the extra points for the ArcAngle segment type.
+
+        It also auto evaluates the arc_plane if it was not specified by the user.
+        segment.extra_points[0] contains the arc mid point (on the arc).
+        segment.extra_points[1] contains the arc end point.
+        Both are function of the arc center, arc angle and arc plane.
+        """
+        # from start-point and angle, calculate the mid-point and end-points
+        # Also identify the plane of the arc ("YZ", "ZX", "XY")
+        plane_axes = {"YZ": [1, 2], "ZX": [2, 0], "XY": [0, 1]}
+        c_xyz = self._primitives.value_in_object_units(segment.arc_center)
+        p0_xyz = self._primitives.value_in_object_units(start_point)
+
+        if segment.arc_plane:
+            # Accept the user input for the plane of rotation - let the modeler fail if invalid
+            plane_def = (segment.arc_plane, plane_axes[segment.arc_plane])
+        else:
+            # Compare the numeric values of start-point and center-point to determine the orientation plane
+            if c_xyz[0] == p0_xyz[0]:
+                plane_def = ("YZ", plane_axes["YZ"])
+            elif c_xyz[1] == p0_xyz[1]:
+                plane_def = ("ZX", plane_axes["ZX"])
+            elif c_xyz[2] == p0_xyz[2]:
+                plane_def = ("XY", plane_axes["XY"])
+            else:
+                raise Exception("Start point and arc-center do not lie on a common base plane.")
+            segment.arc_plane = plane_def[0]
+
+        # Calculate the extra two points of the angular arc in the alpha-beta plane
+        alph_index = plane_def[1][0]
+        beta_index = plane_def[1][1]
+        c_alph = c_xyz[alph_index]
+        c_beta = c_xyz[beta_index]
+        p0_alph = p0_xyz[alph_index] - c_alph
+        p0_beta = p0_xyz[beta_index] - c_beta
+
+        # rotate to generate the new points
+        arc_ang = self._primitives._app.evaluate_expression(segment.arc_angle)  # in radians
+        h_arc_ang = arc_ang * 0.5
+
+        p1_alph = c_alph + p0_alph * math.cos(h_arc_ang) - p0_beta * math.sin(h_arc_ang)
+        p1_beta = c_beta + p0_alph * math.sin(h_arc_ang) + p0_beta * math.cos(h_arc_ang)
+        p2_alph = c_alph + p0_alph * math.cos(arc_ang) - p0_beta * math.sin(arc_ang)
+        p2_beta = c_beta + p0_alph * math.sin(arc_ang) + p0_beta * math.cos(arc_ang)
+
+        # Generate the 2 new points in XYZ
+        p1 = list(p0_xyz)
+        p1[alph_index] = p1_alph
+        p1[beta_index] = p1_beta
+        p2 = list(p0_xyz)
+        p2[alph_index] = p2_alph
+        p2[beta_index] = p2_beta
+        segment.extra_points = [p1, p2]
+        return True
+
+    @pyaedt_function_handler()
+    def _segment_array(self, segment_data, start_index=0, start_point=None):
+        """Retrieve a property array for a polyline segment for use in the
+        :class:`ansys.aedt.core.modeler.cad.primitives.Polyline` constructor.
+
+        Parameters
+        ----------
+        segment_data : :class:`ansys.aedt.core.modeler.cad.primitives.PolylineSegment` or str
+            Pointer to the calling object that provides additional functionality
+            or a string with the segment type ``Line`` or ``Arc``.
+        start_index : int, string
+            Starting vertex index of the segment within a compound polyline. The
+            default is ``0``.
+        start_point : list, optional
+            Position of the first point for type ``AngularArc``. The default is
+            ``None``. Float values are considered in model units.
+
+        Returns
+        -------
+        list
+            List of properties defining a polyline segment.
+
+        """
+        if isinstance(segment_data, str):
+            segment_data = PolylineSegment(segment_data)
+
+        seg = [
+            "NAME:PLSegment",
+            "SegmentType:=",
+            segment_data.type,
+            "StartIndex:=",
+            start_index,
+            "NoOfPoints:=",
+            segment_data.num_points,
+        ]
+        if segment_data.type != "Line":
+            seg += ["NoOfSegments:=", f"{segment_data.num_seg}"]
+
+        if segment_data.type == "AngularArc":
+            # from start-point and angle, calculate the mid- and end-points
+            if not start_point:
+                raise ValueError("Start point must be defined for segment of type Angular Arc")
+            self._evaluate_arc_angle_extra_points(segment_data, start_point)
+
+            mod_units = self._primitives.model_units
+            seg += [
+                "ArcAngle:=",
+                segment_data.arc_angle,
+                "ArcCenterX:=",
+                f"{_dim_arg(segment_data.arc_center[0], mod_units)}",
+                "ArcCenterY:=",
+                f"{_dim_arg(segment_data.arc_center[1], mod_units)}",
+                "ArcCenterZ:=",
+                f"{_dim_arg(segment_data.arc_center[2], mod_units)}",
+                "ArcPlane:=",
+                segment_data.arc_plane,
+            ]
+
+        return seg
+
+    @pyaedt_function_handler(abstol="tolerance")
+    def remove_point(self, position, tolerance=1e-9):
+        """Remove a point from an existing polyline by position.
+
+        You must enter the exact position of the vertex as a list
+        of ``[x, y, z]`` coordinates in the object's coordinate system.
+
+        Parameters
+        ----------
+        position : list
+            List of ``[x, y, z]`` coordinates specifying the vertex to remove.
+        tolerance : float, optional
+            Absolute tolerance of the comparison of a specified position to the
+            vertex positions. The default is ``1e-9``.
+
+        Returns
+        -------
+        bool
+            ``True`` when successful, ``False`` when failed.
+
+        References
+        ----------
+        >>> oEditor.DeletePolylinePoint
+
+        Examples
+        --------
+        Use floating point values for the vertex positions.
+
+        >>> P = modeler.create_polyline([[0, 1, 2], [0, 2, 3], [2, 1, 4]])
+        >>> P.remove_point([0, 1, 2])
+
+        Use string expressions for the vertex position.
+
+        >>> P = modeler.create_polyline([[0, 1, 2], [0, 2, 3], [2, 1, 4]])
+        >>> P.remove_point(["0mm", "1mm", "2mm"])
+
+        Use string expressions for the vertex position and include an absolute
+        tolerance when searching for the vertex to be removed.
+
+        >>> P = modeler.create_polyline([[0, 1, 2], [0, 2, 3], [2, 1, 4]])
+        >>> P.remove_point(["0mm", "1mm", "2mm"],tolerance=1e-6)
+        """
+        if not self.is_polyline:
+            self.logger.error("Method remove_point applies only to Polyline objects.")
+            return False
+        found_vertex = False
+        seg_id = None
+        at_start = None
+        pos_xyz = self._primitives.value_in_object_units(position)
+        for ind, point_pos in enumerate(self.points):
+            # compare the specified point with the vertex data using an absolute tolerance
+            # (default of math.isclose is 1e-9 which should be ok in almost all cases)
+            found_vertex = GeometryOperators.points_distance(point_pos, pos_xyz) <= tolerance
+            if found_vertex:
+                if ind == len(self.points) - 1:
+                    at_start = False
+                    seg_id = self._get_segment_id_from_point_n(ind, at_start, allow_inner_points=True)
+                else:
+                    at_start = True
+                    seg_id = self._get_segment_id_from_point_n(ind, at_start, allow_inner_points=True)
+                break
+
+        if not found_vertex or seg_id is None or at_start is None:
+            raise ValueError(f"Specified vertex {position} not found in polyline {self._m_name}.")
+
+        try:
+            self._primitives.oeditor.DeletePolylinePoint(
+                [
+                    "NAME:Delete Point",
+                    "Selections:=",
+                    self._m_name + ":CreatePolyline:1",
+                    "Segment Indices:=",
+                    [seg_id],
+                    "At Start:=",
+                    at_start,
+                ]
+            )
+        except Exception:  # pragma: no cover
+            raise ValueError(f"Invalid edge ID {seg_id} is specified on polyline {self.name}.")
+        else:
+            i_start, i_end = self._get_point_slice_from_segment_id(seg_id, at_start)
+            del self._positions[i_start:i_end]
+            del self._segment_types[seg_id]
+
+        return True
+
+    @pyaedt_function_handler(segment_id="assignment")
+    def remove_segments(self, assignment):
+        """Remove a segment from an existing polyline by segment id.
+
+        You must enter the segment id or the list of the segment ids you want to remove.
+
+        Parameters
+        ----------
+        assignment : int or List of int
+            One or more edge IDs within the total number of edges of the polyline.
+
+        Returns
+        -------
+        bool
+            ``True`` when successful, ``False`` when failed.
+
+        References
+        ----------
+        >>> oEditor.DeletePolylinePoint
+
+        Examples
+        --------
+        >>> P = modeler.create_polyline([[0, 1, 2], [0, 2, 3], [2, 1, 4]])
+        >>> P.remove_segments(assignment=0)
+        """
+        if not self.is_polyline:
+            self.logger.error("Method remove_point applies only to Polyline objects.")
+            return False
+        if isinstance(assignment, int):
+            assignment = [assignment]
+        elif isinstance(assignment, list):
+            assignment.sort()
+        else:
+            raise TypeError("segment_id must be int or list of int.")
+        try:
+            self._primitives.oeditor.DeletePolylinePoint(
+                [
+                    "NAME:Delete Point",
+                    "Selections:=",
+                    self.name + ":CreatePolyline:1",
+                    "Segment Indices:=",
+                    assignment,
+                    "At Start:=",
+                    True,
+                ]
+            )
+        except Exception:  # pragma: no cover
+            raise ValueError(f"Invalid segment ID {assignment} is specified on polyline {self.name}.")
+        else:
+            assignment.reverse()
+            for sid in assignment:
+                if sid == len(self._segment_types) - 1:
+                    # removing the last segment, AEDT removes ALWAYS the last polyline point
+                    at_start = False
+                else:
+                    at_start = True
+                i_start, i_end = self._get_point_slice_from_segment_id(sid, at_start)
+                del self._positions[i_start:i_end]
+                del self._segment_types[sid]
+        return True
+
+    @pyaedt_function_handler(type="section")
+    def set_crosssection_properties(
+        self, section=None, orient=None, width=0, topwidth=0, height=0, num_seg=0, bend_type=None
+    ):
+        """Set the properties of an existing polyline object.
+
+        Parameters
+        ----------
+        section : str, optional
+            Types of the cross-sections. Options are ``"Line"``, ``"Circle"``, ``"Rectangle"``,
+            and ``"Isosceles Trapezoid"``. The default is ``None``.
+        orient : str, optional
+            Direction of the normal vector to the width of the cross-section.
+            Options are ``"X"``, ``"Y"``, ``"Z"``, and ``"Auto"``. The default
+            is ``None``, which sets the orientation to ``"Auto"``.
+        width : float or str, optional
+           Width or diameter of the cross-section for all types. The default is
+           ``0``.
+        topwidth : float or str
+           Top width of the cross-section for the type ``"Isosceles Trapezoid"``
+           only. The default is ``0``.
+        height : float or str
+            Height of the cross-section for the types ``"Rectangle"`` and `"Isosceles
+            Trapezoid"`` only. The default is ``0``.
+        num_seg : int, optional
+            Number of segments in the cross-section surface for the types ``"Circle"``,
+            ``"Rectangle"``, and ``"Isosceles Trapezoid"``. The default is ``0``.
+            The value must be ``0`` or greater than ``2``.
+        bend_type : str, optional
+            Type of the bend. The default is ``None``, in which case the bend type
+            is set to ``"Corner"``. For the type ``"Circle"``, the bend type should be
+            set to ``"Curved"``.
+
+        Returns
+        -------
+        bool
+            ``True`` when successful, ``False`` when failed.
+
+        References
+        ----------
+        >>> oEditor.ChangeProperty
+
+        Examples
+        --------
+        >>> P = modeler.create_polyline([[0, 1, 2], [0, 2, 3], [2, 1, 4]])
+        >>> P.set_crosssection_properties(section="Circle",width="1mm")
+
+        """
+        if not self.is_polyline:
+            self.logger.error("Method remove_point applies only to Polyline objects.")
+            return False
+        # Set the default section type to "None"
+        section_type = section
+        if not section_type:
+            section_type = "None"
+
+        # Set the default orientation to "Auto"
+        section_orient = orient
+        if not section_orient:
+            section_orient = "Auto"
+
+        # Set the default bend-type to "Corner"
+        section_bend = bend_type
+        if not section_bend:
+            section_bend = "Corner"
+
+        # Ensure number-of segments is valid
+        if num_seg and num_seg < 3:
+            raise ValueError("Number of segments for a cross-section must be 0 or greater than 2.")
+
+        model_units = self._primitives.model_units
+
+        arg1 = ["NAME:AllTabs"]
+        arg2 = ["NAME:Geometry3DCmdTab", ["NAME:PropServers", self._m_name + ":CreatePolyline:1"]]
+        arg3 = ["NAME:ChangedProps"]
+        arg3.append(["NAME:Type", "Value:=", section_type])
+        arg3.append(["NAME:Orientation", "Value:=", section_orient])
+        arg3.append(["NAME:Bend Type", "Value:=", section_bend])
+        arg3.append(["NAME:Width/Diameter", "Value:=", _dim_arg(width, model_units)])
+        if section_type == "Rectangle":
+            arg3.append(["NAME:Height", "Value:=", _dim_arg(height, model_units)])
+        elif section_type == "Circle":
+            arg3.append(["NAME:Number of Segments", "Value:=", num_seg])
+        elif section_type == "Isosceles Trapezoid":
+            arg3.append(["NAME:Top Width", "Value:=", _dim_arg(topwidth, model_units)])
+            arg3.append(["NAME:Height", "Value:=", _dim_arg(height, model_units)])
+        arg2.append(arg3)
+        arg1.append(arg2)
+        self._primitives.oeditor.ChangeProperty(arg1)
+        return True
+
+    @pyaedt_function_handler()
+    def _get_point_slice_from_segment_id(self, segment_id, at_start=True):
+        """Get the points belonging to the segment from the segment id.
+
+        The points are returned as list slice by returning the indexes.
+
+        Parameters
+        ----------
+        segment_id : int
+            Segment id.
+
+        at_start : bool
+            if ``True`` the slice includes the start point of the segment and not the end point.
+            If ``False`` the slice includes the end point of the segment and not the start point.
+
+        Returns
+        -------
+        tuple of int, bool
+            Indexes of the list slice. ``False`` when failed.
+        """
+        i_end = 0
+        for i, s in enumerate(self.segment_types):
+            i_start = i_end
+            if s.type == "Line":
+                i_end += 1
+            elif s.type == "Arc":
+                i_end += 2
+            elif s.type == "AngularArc":
+                i_end += 2
+            elif s.type == "Spline":
+                i_end += s.num_points - 1
+            if i == segment_id:
+                if at_start:
+                    return i_start, i_end
+                else:
+                    return i_start + 1, i_end + 1
+        return False
+
+    @pyaedt_function_handler()
+    def _get_segment_id_from_point_n(self, pn, at_start, allow_inner_points=False):
+        """Get the segment id for a given point index considering the segment types in the polyline.
+
+        If a segment cannot be found with the specified rules, the function returns False.
+
+        Parameters
+        ----------
+        pn : int
+            Point number along the polyline.
+        at_start : bool
+            If set to ``True`` the segment id that begins with the point pn is returned.
+            If set to ``False`` the segment id that terminates with the point pn is returned.
+        allow_inner_points : bool, optional
+            If set to ``False`` only points that are at the extremities of the segments are considered.
+            If pn is in the middle of a segment, the function returns False.
+            If set to ``True`` also points in the middle of the segments are considered.
+
+        Returns
+        -------
+        int, bool
+            Segment id when successful. ``False`` when failed.
+        """
+        n_points = 0
+        for i, s in enumerate(self.segment_types):
+            if n_points == pn and at_start:
+                return i
+            n_points_imu = n_points
+            if s.type == "Line":
+                n_points += 1
+            elif s.type == "Arc":
+                n_points += 2
+            elif s.type == "AngularArc":
+                n_points += 2
+            elif s.type == "Spline":
+                n_points += s.num_points - 1
+            if n_points == pn and not at_start:
+                return i
+            if n_points_imu < pn < n_points and allow_inner_points:
+                return i
+        return False
+
+    @pyaedt_function_handler(position_list="points")
+    def insert_segment(self, points, segment=None):
+        """Add a segment to an existing polyline.
+
+        Parameters
+        ----------
+        points : List
+            List of positions of the points that define the segment to insert.
+            Either the starting point or ending point of the segment list must
+            match one of the vertices of the existing polyline.
+        segment : str or :class:`ansys.aedt.core.modeler.cad.primitives.PolylineSegment`, optional
+            Definition of the segment to insert. For the types ``"Line"`` and ``"Arc"``,
+            use their string values ``"Line"`` and ``"Arc"``. For the types ``"AngularArc"``
+            and ``"Spline"``, use the :class:`ansys.aedt.core.modeler.cad.primitives.PolylineSegment`
+            object to define the segment precisely. The default is ``None``.
+
+        Returns
+        -------
+        bool
+            ``True`` when successful, ``False`` when failed.
+
+        References
+        ----------
+        >>> oEditor.InsertPolylineSegment
+
+        """
+        # Check for a valid number of points
+        if not self.is_polyline:
+            self.logger.error("Method remove_point applies only to Polyline objects.")
+            return False
+        num_points = len(points)
+
+        # define the segment type from the number of points given
+        if not segment:
+            if num_points < 2:
+                raise ValueError("num_points must contain at least 2 points to auto-define a segment.")
+            if num_points == 2:
+                segment = PolylineSegment("Line")
+            elif num_points == 3:
+                segment = PolylineSegment("Arc")
+            else:  # num_points>3
+                segment = PolylineSegment("Spline", num_points=num_points)
+        else:
+            if isinstance(segment, str) and segment in ["Line", "Arc"]:
+                segment = PolylineSegment(segment)
+                num_points = segment.num_points
+            elif isinstance(segment, PolylineSegment):
+                num_points = segment.num_points
+                if segment.type == "AngularArc":
+                    self._evaluate_arc_angle_extra_points(segment, start_point=points[0])
+            else:
+                raise TypeError('segment must be either "Line", "Arc" or PolylineSegment object.')
+            if segment.type != "AngularArc" and len(points) < num_points:
+                raise ValueError("position_list must contain enough points for the specified segment type.")
+            elif segment.type == "AngularArc" and len(points) < 1:
+                raise ValueError("position_list must contain the start point for AngularArc segment.")
+
+        # Check whether start-point and end-point of the segment is in the existing polylines points
+        start_point = points[0]
+
+        # End point does not exist for an AngularArc
+        if segment.type != "AngularArc":
+            end_point = points[-1]
+        else:
+            end_point = []
+
+        at_start = None
+        p_insert_position = None
+        insert_points = None
+        num_polyline_points = len(self.points)
+        i = None
+        for i, point in enumerate(self.points):
+            if end_point and (
+                GeometryOperators.points_distance(
+                    self._primitives.value_in_object_units(point), self._primitives.value_in_object_units(end_point)
+                )
+                < 1e-8
+            ):
+                at_start = True
+                p_insert_position = i
+                insert_points = points[: num_points - 1]  # All points but last one.
+                if i == num_polyline_points - 1:
+                    if segment.type != "Line":
+                        # Inserting a segment in this position is not allowed in AEDT.
+                        # We can make it work only for "Line" segments.
+                        return False
+                    at_start = False
+                    points = [self.points[-2], start_point]
+                break
+            elif (
+                GeometryOperators.points_distance(
+                    self._primitives.value_in_object_units(point), self._primitives.value_in_object_units(start_point)
+                )
+                < 1e-8
+            ):
+                # note that AngularArc can only be here
+                at_start = False
+                p_insert_position = i + 1
+                if segment.type != "AngularArc":
+                    insert_points = points[1:num_points]  # Insert all points but first one
+                else:
+                    insert_points = segment.extra_points[:]  # For AngularArc insert the extra points
+                if i == 0:
+                    if segment.type != "Line":
+                        # Inserting a segment in this position is not allowed in AEDT.
+                        # PyAEDT can make it work only for "Line" segments.
+                        return False
+                    at_start = True
+                    points = [end_point, self.points[1]]
+                break
+
+        if p_insert_position is None or insert_points is None:
+            raise RuntimeError("Point for the insert is not found.")
+
+        if i is None:
+            raise ValueError("The polyline contains no points. It is impossible to insert a segment.")
+        segment_index = self._get_segment_id_from_point_n(i, at_start=at_start)
+
+        if not isinstance(segment_index, int):
+            raise ValueError("Segment for the insert is not found.")
+        if at_start:
+            s_insert_position = segment_index
+        else:
+            s_insert_position = segment_index + 1
+        segment_type = segment.type
+
+        arg_1 = [
+            "NAME:Insert Polyline Segment",
+            "Selections:=",
+            self._m_name + ":CreatePolyline:1",
+            "Segment Indices:=",
+            [segment_index],
+            "At Start:=",
+            at_start,
+            "SegmentType:=",
+            segment_type,
+        ]
+
+        # Points and segment data
+        arg_2 = ["NAME:PolylinePoints"]
+        if segment.type in ("Line", "Spline", "Arc"):
+            for pt in points[0:num_points]:
+                arg_2.append(self._pl_point(pt))
+            arg_1.append(arg_2)
+        elif segment.type == "AngularArc":
+            seg_str = self._segment_array(segment, start_point=start_point)
+            arg_2.append(self._pl_point(start_point))
+            arg_2.append(self._pl_point(segment.extra_points[0]))
+            arg_2.append(self._pl_point(segment.extra_points[1]))
+            arg_1.append(arg_2)
+            arg_1 += seg_str[9:]
+        self._primitives.oeditor.InsertPolylineSegment(arg_1)
+
+        # check if the polyline has been modified correctly
+        if self._check_polyline_health() is False:
+            raise ValueError("Adding the segment result in an unclassified object. Undoing operation.")
+
+        # add the points and the segment to the object
+        self._positions[p_insert_position:p_insert_position] = insert_points
+        self._segment_types[s_insert_position:s_insert_position] = [segment]
+
+        return True
+
+    @pyaedt_function_handler()
+    def _check_polyline_health(self):
+        # force re-evaluation of object_type
+        self._object_type = None
+        if self.object_type == "Unclassified":
+            # Undo operation
+            self._primitives._app.odesign.Undo()
+            self._object_type = None
+            if self.object_type == "Unclassified":
+                raise RuntimeError("Undo operation failed.")
+            return False
+        return True
+
+
+class PolylineSegment(object):
+    """Creates and manipulates a segment of a polyline.
+
+    Parameters
+    ----------
+    segment_type : str
+        Type of the object. Choices are ``"Line"``, ``"Arc"``, ``"Spline"``,
+        and ``"AngularArc"``.
+    num_seg : int, optional
+        Number of segments for the types ``"Arc"``, ``"Spline"``, and
+        ``"AngularArc"``.  The default is ``0``. For the type
+        ``Line``, this parameter is ignored.
+    num_points : int, optional
+        Number of control points for the type ``Spline``. For other
+        types, this parameter is defined automatically.
+    arc_angle : float or str, optional
+        Sweep angle in radians or a valid value string. For example,
+        ``"35deg"`` or ``0.25``.
+        This argument is Specific to type AngularArc.
+    arc_center : list or str, optional
+        List of values in model units or a valid value string. For
+        example, a list of ``[x, y, z]`` coordinates.
+        This argument is Specific to type AngularArc.
+    arc_plane : str, int optional
+        Plane in which the arc sweep is performed in the active
+        coordinate system ``"XY"``, ``"YZ"`` or ``"ZX"``. The default is
+        ``None``, in which case the plane is determined automatically
+        by the first coordinate for which the starting point and
+        center point have the same value.
+        This argument is Specific to type AngularArc.
+
+    Examples
+    --------
+    See :class:`ansys.aedt.core.Primitives.Polyline`.
+
+    """
+
+    def __init__(self, segment_type, num_seg=0, num_points=0, arc_angle=0, arc_center=None, arc_plane=None):
+        valid_types = ["Line", "Arc", "Spline", "AngularArc"]
+        if segment_type not in valid_types:
+            raise TypeError(f"Segment type must be one of {valid_types}.")
+        self.type = segment_type
+        if segment_type != "Line":
+            self.num_seg = num_seg
+        if segment_type == "Line":
+            self.num_points = 2
+        if segment_type == "Spline":
+            self.num_points = num_points
+        if "Arc" in segment_type:
+            self.num_points = 3
+        if segment_type == "AngularArc":
+            self.arc_angle = arc_angle
+            if not arc_center:
+                arc_center = [0, 0, 0]
+            if len(arc_center) != 3:
+                raise ValueError("Arc center must be a list of length 3.")
+            self.arc_center = arc_center
+        if isinstance(arc_plane, int):
+            if arc_plane == PLANE.XY:
+                self.arc_plane = "XY"
+            elif arc_plane == PLANE.ZX:
+                self.arc_plane = "ZX"
+            elif arc_plane == PLANE.YZ:
+                self.arc_plane = "YZ"
+            else:
+                raise ValueError("arc_plane must be 0, 1, or 2 ")
+        elif arc_plane:
+            if arc_plane not in ["XY", "ZX", "YZ"]:
+                raise ValueError('arc_plane must be "XY", "ZX", or "YZ" ')
+            self.arc_plane = arc_plane
+        else:
+            self.arc_plane = None
+        self.extra_points = None
