+++ tests/system/general/test_08_Primitives3D.py
@@ -442,7 +442,7 @@ class TestClass:
         assert P.is3d == False
         assert isinstance(P.color, tuple)
         get_P = self.aedtapp.modeler["Poly1"]
-        assert isinstance(get_P, Polyline)
+        assert isinstance(get_P, Object3d)
         P2 = self.aedtapp.modeler.create_polyline(
             arrofpos, cover_surface=False, name="Poly_nonmodel", material="Copper", non_model=True
         )
@@ -913,12 +913,12 @@ class TestClass:
         P3 = P.clone()
         P4 = P.clone()
 
-        P1.set_crosssection_properties(type="Line", width="1mm")
+        P1.set_crosssection_properties(section="Line", width="1mm")
         a = P1.object_type
 
-        P2.set_crosssection_properties(type="Circle", width="1mm", num_seg=5)
-        P3.set_crosssection_properties(type="Rectangle", width="1mm", height="1mm")
-        P4.set_crosssection_properties(type="Isosceles Trapezoid", width="1mm", height="1mm", topwidth="4mm")
+        P2.set_crosssection_properties(section="Circle", width="1mm", num_seg=5)
+        P3.set_crosssection_properties(section="Rectangle", width="1mm", height="1mm")
+        P4.set_crosssection_properties(section="Isosceles Trapezoid", width="1mm", topwidth="4mm", height="1mm")
 
         assert P.object_type == "Line"
         assert P1.object_type == "Sheet"
@@ -950,7 +950,7 @@ class TestClass:
     def test_51_remove_edges_from_polyline(self):
         modeler = self.aedtapp.modeler
         P = modeler.create_polyline([[0, 1, 2], [0, 2, 3], [2, 1, 4]])
-        P.remove_edges(assignment=0)
+        P.remove_segments(assignment=0)
         assert P.points == [[0, 2, 3], [2, 1, 4]]
         assert len(P.segment_types) == 1
         assert P.name in self.aedtapp.modeler.line_names
@@ -982,7 +982,7 @@ class TestClass:
 
     def test_52_remove_edges_from_polyline_invalid(self):
         P = self.aedtapp.modeler.create_polyline([[0, 1, 2], [0, 2, 3], [2, 1, 4]])
-        P.remove_edges(assignment=[0, 1])
+        P.remove_segments(assignment=[0, 1])
         assert not P.name in self.aedtapp.modeler.line_names
 
     def test_53_duplicate_polyline_and_manipulate(self):
@@ -1008,7 +1008,7 @@ class TestClass:
             name="Inductor1",
         )
 
-        ind.set_crosssection_properties(type="Circle", width=wireThickness_um)
+        ind.set_crosssection_properties(section="Circle", width=wireThickness_um)
 
         polyline_points = ind.points
